materials-engine/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
pydantic>=2.5.0
pyyaml>=6.0
pytest>=7.0.0
pymatgen>=2024.0.0
